requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==3.0.0
asyncpg==0.30.0
attrs==25.1.0
Automat==24.8.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
constantly==23.10.4
cryptography==44.0.1
cssselect==1.2.0
dataclasses-json==0.6.7
decorator==5.2.1
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
dydantic==0.0.8
email_validator==2.2.0
executing==2.2.0
fake-useragent==2.0.3
fastapi==0.115.10
fastapi-cli==0.0.7
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
filelock==3.17.0
frozenlist==1.5.0
gotrue==2.11.4
gpt4all==2.8.2
greenlet==3.1.1
groq==0.18.0
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
hyperframe==6.1.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipython==9.0.1
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.19
langchain-anthropic==0.3.9
langchain-community==0.3.18
langchain-core==0.3.41
langchain-groq==0.2.4
langchain-nomic==0.1.4
langchain-openai==0.3.7
langchain-text-splitters==0.3.6
langgraph==0.3.2
langgraph-checkpoint==2.0.16
langgraph-prebuilt==0.1.1
langgraph-sdk==0.1.53
langmem==0.0.14
langsmith==0.3.11
loguru==0.7.3
lxml==5.3.1
makefun==1.15.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
mdurl==0.1.2
msgpack==1.1.0
multidict==6.1.0
mypy-extensions==1.0.0
nomic==3.4.1
numpy==2.2.3
openai==1.65.3
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
passlib==1.7.4
pexpect==4.9.0
pillow==10.4.0
playwright==1.50.0
postgrest==0.19.3
prompt_toolkit==3.0.50
propcache==0.3.0
Protego==0.4.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg2==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pwdlib==0.2.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
queuelib==1.7.0
realtime==2.4.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
rich==13.9.4
rich-toolkit==0.13.2
schedule==1.2.2
scrapinghub==2.5.0
Scrapy==2.12.0
scrapy-playwright==0.0.43
scrapy-rotating-proxies==0.6.2
scrapy-user-agents==0.1.1
scrapyd==1.5.0
scrapyd-client==2.0.1
service-identity==24.2.0
setuptools==75.8.2
shellingham==1.5.4
shub==2.15.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.38
stack-data==0.6.3
starlette==0.46.0
storage3==0.11.3
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
tenacity==9.0.0
tiktoken==0.9.0
tldextract==5.1.3
toml==0.10.2
tqdm==4.55.1
traitlets==5.14.3
trustcall==0.0.37
Twisted==24.11.0
typer==0.15.2
typing==3.10.0.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.3
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
uberegg==0.1.1
urllib3==2.3.0
user-agents==2.2.0
uvicorn==0.34.0
uvloop==0.21.0
w3lib==2.3.1
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
